// client/src/store/types/action-types.ts
export enum Actions {
  UPDATE_USERINFO = 'UPDATE_USERINFO',
  UPDATE_IS_AUTH = 'UPDATE_IS_AUT',
  UPDATE_SEARCH_GAME = 'GETTING_SEARCH_GAME',
  GETTING_GENRES = 'GETTING_GENRES',
  UPDATE_GAMES_ON_GENRES = 'GETTING_GAMES_ON_GENRES',
  GETTING_THEMES = 'GETTING_THEMES',
  UPDATE_GAMES_ON_THEMES = 'UPDATE_GAMES_ON_THEMES',
  GETTING_PLATFORMS = 'GETTING_PLATFORMS',
  UPDATE_GAMES_ON_PLATFORMS = 'UPDATE_GAMES_ON_PLATFORMS',
  UPDATE_BASKET = 'UPDATE_BASKET',
  REMOVE_BASKET = 'REMOVE_BASKET',
  GETTING_MAIN_GAMES = 'GETTING_MAIN_GAMES',
}